refactor: myDate 디렉터리 구조 변경
helper 디렉터리로 이동
